requirements.txt
annotated-types==0.6.0
anyio==4.2.0
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.9.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
lxml==4.9.4
openai==1.6.1
pip==23.3.1
pydantic-core==2.14.6
pydantic==2.5.3
python-dateutil==2.8.2
python-dotenv~=0.21.1
pytz==2023.3.post1
requests==2.31.0
setuptools==68.2.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
tqdm==4.66.1
typing-extensions==4.9.0
urllib3==2.0.7
wheel==0.41.2
boto3==1.34.20
botocore==1.34.20
pelican~=4.9.1